chore: Restructured directory hierarchy.
