+++ blocks/adaptors/df_numpy.py
