+++ blocks/models/base.py
@@ -1,12 +1,7 @@
 from type import BaseConfig
-from abc import ABC
 import pandas as pd
 from typing import Optional
-from runner.store import Store
-
-
-class Block(ABC):
-    id: str
+from blocks.base import Block
 
 
 class Model(Block):
@@ -27,17 +22,3 @@ class Model(Block):
 
     def is_fitted(self) -> bool:
         raise NotImplementedError()
-
-
-class DataSource(Block):
-
-    id: str
-
-    def __init__(self, id: str):
-        self.id = id
-
-    def deplate(self, store: Store) -> pd.DataFrame:
-        return store.get_data(self.id)
-
-    def preload(self):
-        pass
