+++ blocks/models/pytorch/decoder.py
@@ -1,7 +1,6 @@
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
-from model.base import Block
 from type import PytorchConfig
 
 device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
