+++ library/__init__.py
@@ -1,4 +1,4 @@
-from .examples.simple_pipeline import simple_pipeline
-from .examples.text_image_pipeline import text_image_pipeline
-from .examples.simple_decoder import all_transformations
+# from .examples.simple_pipeline import simple_pipeline
+# from .examples.text_image_pipeline import text_image_pipeline
+from .examples.all_transformations import all_transformations
 from .examples.hate_speech import hate_speech_detection_pipeline
