+++ run_pipeline.py
@@ -6,11 +6,12 @@ from library.examples.hate_speech import (
     hate_speech_detection_pipeline,
     preprocess_config,
 )
+from library.examples.all_transformations import all_transformations
 
 train_dataset, test_dataset = load_data("data/original", preprocess_config)
 
 train_pipeline(
-    hate_speech_detection_pipeline(),
+    all_transformations(),
     {"input": train_dataset},
     train_dataset[Const.label_col],
 )
